Created home page to check for user sign in and redirects to other pages
